Fix default positioning strategy handling
that causes flaky tests, because default positioning strategy was
dependent on the order of the reported messages.
The code lead to it was introduced in an attempt to extract common
PSI-independent strategy, because PSI is leaking into the abstract
diagnostic infrastructure. The approach is definitely problematic,
but to fix it now, the leaking dependency to the psi-based module
is introduced. This should be fixed in the future by introducing
better abstractions.
Fixes flaky tests touched in the commit.