[LL] Remove .kjsm library symbol providers
It's not clear whether the provider is necessary for K2 IDE and whether
there will be libraries with .kjsm files that we are going to support.
Since there are no tests that can prove the correct work of the symbol
provider right now, it's better to not add it now and return later if
needed. The existing tests cover only .knm files.

KTIJ-27566